pyproject.toml
[tool.poetry]
name = "langchain-docs"
version = "0.1.0"
description = ""
authors = ["Markiyan Pyts <markiyan.pyts@osf.digital>"]
readme = "README.md"
packages = [{include = "langchain_docs"}]

[tool.poetry.dependencies]
python = "^3.11"
bs4 = "^0.0.1"
requests = "^2.31.0"
langchain = "^0.0.228"
beautifulsoup4 = "^4.12.2"
black = "^23.3.0"
tiktoken = "^0.4.0"
openai = "^0.27.8"
pinecone-client = "^2.2.2"
nltk = "^3.8.1"
fastapi = "^0.100.0"
jinja2 = "^3.1.2"
uvicorn = "^0.22.0"
streamlit = "^1.24.1"
streamlit-chat = "^0.1.1"
tqdm = "^4.65.0"
lxml = "^4.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
